bug dashboard corrigé

+++ controllers/dashboard/galleries/delete-ctrl.php
@@ -10,24 +10,39 @@ $auth = Auth::check();
 
 
 try {
+    $error = [];
+
     // * get the id_gallery to delete
     $id_gallery = intval(filter_input(INPUT_GET, 'id_gallery', FILTER_SANITIZE_NUMBER_INT));
 
-    $delete = Gallery::delete($id_gallery);
 
-    if ($delete) {
+    $gallery = Gallery::get($id_gallery);
+    // dd($gallery);
+    if (isset($gallery->picture_photoCover)) {
+        $error['delete'] = true;
+    }
+    if ($error == []) {
+        // dd($error);
+        $delete = Gallery::delete($id_gallery);
 
-        $msg = 'La donnée a bien été supprimée';
+        if ($delete) {
+
+            $msg = 'La donnée a bien été supprimée !';
+            $_SESSION['msg'] = $msg;
+
+
+            header('Location: /controllers/dashboard/galleries/archive-ctrl.php');
+            die;
+        } else {
+            $msg = 'La donnée n\'a pas été supprimée !';
+        }
+    } else {
+        $msg = "La galerie ne doit plus avoir de photos ni de commentaires pour être supprimée !";
         $_SESSION['msg'] = $msg;
-
 
         header('Location: /controllers/dashboard/galleries/archive-ctrl.php');
         die;
-
-    } else {
-        $msg = 'La donnée n\'a pas été supprimée !';
     }
-
 } catch (\Throwable $th) {
     //throw $th;
 }
